requirements.txt
asyncio
speechrecognition
torch
pyserial
RPi.GPIO